requirements.txt
asyncio
websockets
pandas
sqlalchemy
psutil
pandasql
joblib
logging
uuid
base64
multiprocessing
python-dotenv
flask
pyarrow